16e496b023c742d4a89ecb63d8ffff8362c1263d Make Error non-exhaustive (Christian Lewe)
18f6e30858e8490a69c81bd286d58a9812de6361 Test key lock (Christian Lewe)
c51b914d6d5faf621fc6b45178d5a2e3a66a4c4f Test timelocks (Christian Lewe)
9e6381d84fd06d5d25d043597c4a1e83253d00b3 Change relative height to u16 (Christian Lewe)
97b689028922857efa714426fa4fbbb7ca85c81d Extend dummy environment (Christian Lewe)

Pull request description:

  Correctly implements and tests timelocks. Also tests key locks.

ACKs for top commit:
  apoelstra:
    ACK 16e496b023c742d4a89ecb63d8ffff8362c1263d

Tree-SHA512: 6c7f2ecccd5b235a305543c0c4ff9b75e8dfe866361182c6e2cca1a7ba16b058c59747c24d182390fafdf5eebca188728d27b3d62ee57f002bc39a78fddd243d

+++ src/policy/ast.rs
@@ -48,7 +48,7 @@ pub enum Policy<Pk: MiniscriptKey> {
     /// Absolute timelock
     After(u32),
     /// Relative timelock
-    Older(u32),
+    Older(u16),
     /// Provide the preimage of the given SHA256 hash image
     Sha256(Pk::Sha256),
     /// Satisfy all of the given sub-policies
